third_party/blink/web_tests/media/media-playback-while-not-visible-permission-policy-viewport.html

<!DOCTYPE html>
<title>Test scroll behavior of HTMLMediaElements with the media-playback-while-not-rendered permission policy</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<body>
  <div id="container" style="height: 500px; overflow: scroll;">
    <iframe id="video-iframe" allow="media-playback-while-not-visible 'none'; autoplay *"
        srcdoc="<!DOCTYPE html><video autoplay loop></video>"></iframe>
    <div id="very-long-content" style="height: 5000px;"></div>
  </div>
<script>
  internals.settings.setAutoplayPolicy('no-user-gesture-required');

  async_test(function(t) {
    let iframe = document.getElementById('video-iframe');
    iframe.onload = t.step_func(function() {
      const media = iframe.contentDocument.querySelector('video');
      assert_true(media.paused);

      media.addEventListener('playing', t.step_func(function() {
        assert_false(media.paused);

        // Scroll the iframe out of view.
        let container = document.getElementById('container');
        container.scrollTop = container.scrollHeight;

        // Media should keep playing — scrolling out of viewport does not
        // count as "not rendered".
        setTimeout(t.step_func_done(function() {
          assert_false(media.paused);
        }), 2000);
      }));

      media.src = 'content/test.ogv';
    });
  }, "Media playback should not pause if iframe is scrolled out of view when permission policy is set");
</script>
</body>
